+++ end_to_end_test/lib/values.dart
@@ -503,21 +503,6 @@ abstract class ValueWithOnSet
   ValueWithOnSet._();
 }
 
-abstract class SerializesNullsValue
-    implements Built<SerializesNullsValue, SerializesNullsValueBuilder> {
-  @BuiltValueSerializer(serializeNulls: true)
-  static Serializer<SerializesNullsValue> get serializer =>
-      _$serializesNullsValueSerializer;
-
-  @nullable
-  String get value;
-
-  factory SerializesNullsValue(
-          [void Function(SerializesNullsValueBuilder) updates]) =
-      _$SerializesNullsValue;
-  SerializesNullsValue._();
-}
-
 // Check that `toString` from a mixin is not redefined.
 abstract class CustomToStringValue extends Object
     with CustomToString